HARDCORE FANO 3.0
This is a major refactoring of the Hardcore Fano
project. I *think* there is a bug in Hardcore Fano 2.0
which affects the results. Here, I want to recode
"everything", introducing tests along the way to make sure
everything is as it should be. Also, this is an opportunity
to clean up a bit all the useless junk now floating in
2.0.